refactor example to add wasm
