+++ examples/diagram/calculator/Cargo.toml
@@ -16,6 +16,7 @@ serde_json = { workspace = true }
 schemars = { workspace = true }
 tokio = { version = "1.45.1", features = ["macros", "rt-multi-thread"] }
 tracing-subscriber = "0.3.19"
+calculator_lib = { version = "0.1.0", path = "../calculator_lib" }
 
 [dev-dependencies]
 assert_cmd = "2.0.16"
