+++ examples/diagram/calculator_lib/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "calculator_lib"
+version = "0.1.0"
+edition = "2021"
+description = "calculator lib for other examples"
+
+[dependencies]
+bevy_app = "0.12"
+bevy_core = "0.12"
+bevy_impulse = { version = "0.0.2", path = "../../..", features = ["diagram"] }
+bevy_time = "0.12"
+serde = { workspace = true }
+serde_json = { workspace = true }
+schemars = { workspace = true }
+tracing-subscriber = "0.3.19"
