+++ examples/diagram/calculator_wasm/README.md
@@ -1,0 +1,18 @@
+## calculator_wasm
+
+This is an example of how to build a wasm payload to be used in the diagram editor.
+
+### Building
+
+Install `wasm-pack`
+
+```bash
+cargo install wasm-pack
+```
+
+Build with `wasm-pack`
+
+```bash
+wasm-pack build
+```
+
